Override train-provided defaults for ssh
Train provides defaults for a subset of connection configuration
options (such as user and port).  If there are values in .ssh/config
for these things, they are not used.

This change makes it so that when ssh configuration exists for a host
we will prefer those defaults to the ones that train provided
if the operator hasn't explicitly provided them on the CLI.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>